(PC-18595) feat(homeAPI): move tests next to their component/hook
